requirement.txt
asgiref==3.4.1
certifi==2021.5.30
dj-database-url==2.1.0
Django==3.2.23
django-cors-headers==3.10.1
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
importlib-metadata==4.8.3
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy @ file:///opt/concourse/worker/volumes/live/e6a5a904-ea21-4841-4eec-d53e1ac1014c/volume/numpy_and_numpy_base_1603479626870/work
packaging==21.3
pandas==1.1.5
psycopg2==2.7.7
pyparsing==3.1.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
scikit-learn==0.24.2
scipy==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
sqlparse==0.4.4
threadpoolctl==3.1.0
typing_extensions==4.1.1
whitenoise==5.3.0
zipp==3.6.0
